change img source and change del file condition

+++ util/file.js
@@ -1,7 +1,9 @@
 const fs = require("fs");
 
 const deleteFile = (filepath) => {
-  fs.unlink(filepath, (err) => {});
+  if (fs.existsSync(filepath)) {
+    fs.unlink(filepath, (err) => {});
+  }
 };
 
 exports.deleteFile = deleteFile;
